+++ packages/browser-wallet/src/popup/pages/Account/AccountSettings/AccountSettings.tsx
@@ -5,6 +5,7 @@ import { useTranslation } from 'react-i18next';
 import { accountSettingsRoutes } from './routes';
 import ConnectedSites from './ConnectedSites';
 import ExportPrivateKey from './ExportPrivateKey';
+import AccountStatement from './AccountStatement';
 
 export function AccountSettings() {
     const { t } = useTranslation('account', { keyPrefix: 'settings' });
@@ -17,6 +18,9 @@ export function AccountSettings() {
             <Link className="account-settings-page__link" to={accountSettingsRoutes.exportPrivateKey}>
                 {t('exportPrivateKey.title')}
             </Link>
+            <Link className="account-settings-page__link" to={accountSettingsRoutes.accountStatement}>
+                {t('accountStatement.title')}
+            </Link>
         </NavList>
     );
 }
@@ -27,6 +31,7 @@ export default function AccountSettingRoutes() {
             <Route index element={<AccountSettings />} />
             <Route path={accountSettingsRoutes.connectedSites} element={<ConnectedSites />} />
             <Route path={accountSettingsRoutes.exportPrivateKey} element={<ExportPrivateKey />} />
+            <Route path={accountSettingsRoutes.accountStatement} element={<AccountStatement />} />
         </Routes>
     );
 }