+++ packages/browser-wallet/src/popup/pages/Account/AccountSettings/routes.ts
@@ -1,4 +1,5 @@
 export const accountSettingsRoutes = {
     connectedSites: 'connected-sites',
     exportPrivateKey: 'export-private-key',
+    accountStatement: 'account-statement',
 };