+++ packages/browser-wallet/src/popup/pages/Account/i18n/da.ts
@@ -35,6 +35,12 @@ const t: typeof en = {
             done: 'Færdig',
             export: 'Eksporter',
         },
+        accountStatement: {
+            title: 'Eksportér transaktionslog',
+            description:
+                'Transaktionslogs for en konto kan genereres og hentes fra CCDScan.io.\n\nCCDScan er en Concordium block explorer, og ved at trykke på knappen nedenfor åbnes et denne side i din browser.',
+            link: 'Gå til CCDScan.io',
+        },
     },
     sendCcd: {
         labels: {