+++ packages/browser-wallet/src/shared/constants/url.ts
@@ -6,6 +6,8 @@ const urls = {
     website: 'https://www.concordium.com',
     supportWebsite: 'https://support.concordium.software',
     documentationWebsite: 'https://developer.concordium.software/en/mainnet/index.html',
+    ccdscanMainnet: 'https://ccdscan.io/',
+    ccdscanTestnet: 'https://testnet.ccdscan.io/',
 };
 
 export default urls;